@@ -1,13 +1,12 @@
 name: future_button
 description: Customizable Material and Cupertino buttons with progress bars. Useful if you want to show a loading indicator when user presses some action, and show the result afterwards.
 version: 0.1.2
-author: kekland <kk.erzhan@gmail.com>
 homepage: https://github.com/kekland/future_button
 repository: https://github.com/kekland/future_button
 issue_tracker: https://github.com/kekland/future_button/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0 <2.0.0"
 
 dependencies:
@@ -15,7 +14,7 @@ dependencies:
     sdk: flutter
 
 dev_dependencies:
-  pedantic: ^1.9.0
+  pedantic: ^1.11.1
   flutter_test:
     sdk: flutter
 
